Handle structural traits more gracefully
